+++ app/src/main/java/com/brandonjamesyoung/levelup/data/ItemRow.kt
@@ -1,0 +1,6 @@
+package com.brandonjamesyoung.levelup.data
+
+data class ItemRow(
+    var item: Item,
+    var selected: Boolean = false
+)