+++ app/src/main/java/com/brandonjamesyoung/levelup/data/SettingsDao.kt
@@ -9,7 +9,7 @@ interface SettingsDao {
     fun observeById(id: Int): Flow<Settings>
 
     @Query("SELECT * FROM Settings WHERE id = :id")
-    fun getById(id: Int): Settings?
+    suspend fun getById(id: Int): Settings?
 
     @Insert
     suspend fun insert(settings: Settings)